services/user/requirements.txt
Flask==0.12.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
psycopg2==2.7.1
Flask-Testing==0.6.2
flask-cors==3.0.2
flask-migrate==2.0.4
flask-bcrypt==0.7.1
pyjwt==1.5.0